062 - switched to arts db

+++ lib/features/near_me_page/cubits/near_by_events_cubit.dart
@@ -37,8 +37,8 @@ class NearbyEventsCubit extends MaybeEmitHydratedCubit<NearbyEventsState> {
   Future<void> loadNearbyEvents({
     required double latitude,
     required double longitude,
-    required int minDistance,
-    required int maxDistance,
+    required double minDistance,
+    required double maxDistance,
     bool onBackground = false,
   }) async {
     if (!onBackground) {
